csc-ca2-infra/requirements.txt
requests
certifi
chardet
Click
idna
itsdangerous
Jinja2
MarkupSafe
stripe
toml
urllib3
Werkzeug
pymysql
# google-cloud
# google-cloud-vision
pyjwt